+++ checkpoint/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+ - Add `metadata()` method for loading with SafeTensors checkpoint format.
+
 ## [0.11.20] - 2025-07-28
 
 ### Added